Checks.yml updated, bug fixes & antivpn update
In this update I've fixed some bugs, changed the checks.yml, because some configurations were bad choosen by me. xD
Oh btw... There is now an autoupdate checker. Every 15 minutes is the plugin checking, if a new version is out. You do not want to have this activated? Just set in the config.rayzs file the 'notifyUpdate' to false.

You have from now on to put your own api-key in the config.rayzs to use the vpn-detection-service. It's easy! Just go to this site here and replace the long text in the config.rayzs with your api-key.